apps/api-gateway/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
pydantic==1.10.13
httpx==0.24.1
redis==5.0.1
prometheus-client==0.17.1
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-httpx==0.42b0
opentelemetry-exporter-otlp==1.21.0
tenacity==8.2.3
python-json-logger==2.0.7
